#402 rename card component
