+++ app/src/views/Home.vue
@@ -15,7 +15,7 @@
 import Vue from "vue";
 import Header from "@/components/layout/MainHeader.vue";
 import Carousel from "@/components/layout/Carousel.vue";
-import ImageCard, { Card } from "@/components/layout/ImageCard.vue";
+import ImageCard, { Card } from "@/components/layout/ImageCards.vue";
 import SearchComponent from "@/components/search/SearchComponent.vue";
 import { mapActions } from "vuex";
